uuaasdfasdr

+++ Evaluator.py
@@ -6,9 +6,8 @@ import numpy as np
 
 def psnr(batch_size, ref_image_set, output_image_set):
     psnr_val = 0
-
-    ref_image_set = np.array(ref_image_set)
-    output_image_set = np.array(output_image_set)
+    ref_image_set = np.array(ref_image_set, dtype=np.uint8)
+    output_image_set = np.array(output_image_set, dtype=np.uint8)
 
     if batch_size == 1:
         return compare_psnr(ref_image_set, output_image_set, data_range=ref_image_set.max() - ref_image_set.min())
@@ -18,15 +17,15 @@ def psnr(batch_size, ref_image_set, output_image_set):
         output_image = output_image_set[i]
         psnr_val += compare_psnr(ref_image, output_image, data_range=ref_image.max() - ref_image.min())
 
-    psnr_val =  np.mean(psnr_val)
+    psnr_val = np.mean(psnr_val)
     return psnr_val
 
 
 def ssim(batch_size, ref_image_set, output_image_set):
     ssim_val = 0
 
-    ref_image_set = np.array(ref_image_set)
-    output_image_set = np.array(output_image_set)
+    ref_image_set = np.array(ref_image_set, dtype=np.uint8)
+    output_image_set = np.array(output_image_set, dtype=np.uint8)
 
     if batch_size == 1:
         return compare_ssim(ref_image_set, output_image_set,
